refactor: extract _accumulate_leaflet_tilt_gradient to bt_utils.py
Moved the tilt gradient accumulation utility to modules/energy/bt_utils.py to reduce core module size. Preserved exact mutation and scratch-buffer semantics. Verified with tilt energy tests.
